Current Junk as of 11/18/2015
The current state of the programmer squads' progress. Pretty much all of
this is just proof of concept, not to mention lifted straight out of
tutorials... Tomorrow I'm going to bring in the Pi and start setting up
the routers and all that. Will add a todo list later. Homework time.